Ok, finally everything has sense
This is the correct sentence, the old changes at this point not has sens
e.

The problem was, you've extra '( )' in your OR clause,  creating thr
ee clauses as this :  if( [clause != null] && [clause < count] || [claus
e > count] ), for that allways evaluates third clause and crash causing
NullPointerException.

The fix, set two clauses as this: if( [clause !=
null] && [ [clause < count] || [clause > count] ] )

Regards Ben